requirements.txt
langchain~=0.1.4
python-dotenv~=1.0.1
pypdf2~=3.0.1
requests~=2.31.0
streamlit==1.30.0
langchain-community==0.0.16
langchain-openai==0.0.5
sentence-transformers==2.4.0
chroma-hnswlib==0.7.3
chromadb==0.3.29
rank-bm25==0.2.2
beautifulsoup4==4.12.3
wikipedia==1.4.0
langchainhub==0.1.14
tabulate==0.9.0
unstructured[pdf,s3]==0.12.4
